autopairs / cmp fix might not be needed?
I'm not sure where I copied and pasted this from originally, but the
readme section that it links to no longer exists. This issue seems
related (https://github.com/hrsh7th/nvim-cmp/issues/780) - it sounds
like if you tab to a completion option, then hit "(", there was (is?) a
bug that would insert an extra paren. I can't replicate, so assume bug
is fixed and removing this hack.